Feat: V2

+++ api/v1/views/users.py
@@ -31,6 +31,8 @@ def get_user_id(user_id):
 @app_views.route('/users/<user_id>', methods=['DELETE'])
 def delete_user(user_id):
     user = storage.get(User, user_id)
+    if user_id is None:
+        abort(404)
     if user is None:
         abort(404)
     user.delete()
